+++ tests/testthat/test-docvars.R
@@ -3,7 +3,8 @@ test_that("make_docvars() works", {
     docvar2 <- quanteda:::make_docvars(3L, docname = c("A", "B", "C"))
     docvar3 <- quanteda:::make_docvars(3L, docname = 1:3)
     docvar4 <- quanteda:::make_docvars(10L)
-    docvar5 <- quanteda:::make_docvars(3L, docname = c("A", "B", "B"))
+    docvar5 <- quanteda:::make_docvars(3L, docname = c("A", "B", "B"),
+                                       unique = FALSE)
 
     expect_equal(dim(docvar1), c(0, 3))
     expect_equal(dim(docvar2), c(3, 3))
@@ -18,14 +19,20 @@ test_that("make_docvars() works", {
     expect_equal(docvar5[["docname_"]], c("A.1", "B.1", "B.2"))
     expect_error(quanteda:::make_docvars(n = "3"))
     expect_error(quanteda:::make_docvars(n = 1.4))
+    expect_error(quanteda:::make_docvars(3L, docname = c("A", "B", "B")))
 
-    docvar4 <- quanteda:::make_docvars(5L, c("A", "A", "B", "B", "C"))
+    docvar4 <- quanteda:::make_docvars(5L, c("A", "A", "B", "B", "C"),
+                                       unique = FALSE)
     expect_equal(docvar4[["docname_"]], c("A.1", "A.2", "B.1", "B.2", "C.1"))
-    docvar5 <- quanteda:::make_docvars(5L, c("A", "B", "B", "A", "C"))
+    expect_equal(docvar4[["segid_"]], c(1, 2, 1, 2, 1))
+    docvar5 <- quanteda:::make_docvars(5L, c("A", "B", "B", "A", "C"),
+                                       unique = FALSE)
     expect_equal(docvar5[["docname_"]], c("A.1", "B.1", "B.2", "A.2", "C.1"))
-    docvar6 <- quanteda:::make_docvars(5L, c("A", "A", "B", "B", "C"), unique = FALSE)
-    expect_equal(docvar6[["docname_"]], c("A", "A", "B", "B", "C"))
-    expect_equal(docvar6[["segid_"]], c(1, 1, 1, 1, 1))
+    expect_equal(docvar5[["segid_"]], c(1, 1, 2, 2, 1))
+    docvar6 <- quanteda:::make_docvars(5L, c("A", "A", "B", "B", "C"), 
+                                       unique = FALSE)
+    expect_equal(docvar6[["docname_"]], c("A.1", "A.2", "B.1", "B.2", "C.1"))
+    expect_equal(docvar6[["segid_"]], c(1, 2, 1, 2, 1))
 
 })