+++ NEWS.md
@@ -6,6 +6,7 @@
 
 * Added `tokens_annotate()` to insert tags to a tokens object using a dictionary.
 * Added `tokenize` argument to `dictionary()` to match dictionary values and tokens more easily in Japanese and Chinese (#2476).
+* Update `corpus_reshape()` to segment paragraphs more accurately (#2468).
 
 # quanteda 4.3.1